# ENH: Use `np.dot` to speedup `spatial.distance.hamming`