@lgeiger 2023-11-25
> Are you planning to make more np.dot changes like this in spatial? If so, it's probably best to do them all in one go rather than piecemeal.

This should be the last one. I just discovered that the same patch also applies to `hamming`.