@lgeiger 2023-11-25
> I once again checked `asv continuous -e -b ".*Xdist.*" main hamming-dot` on both x86_64 Linux and Mac M2 max (ARM). In this case, I found no evidence of performance improvement for the sizes of benchmarks we currently have in our suite.

As far as I can tell this is because the `Xdist` benchmarks measure `spatial.distance.cdist` and `spatial.distance.pdist` which for the case of `hamming` calls a compiled C function which this PR doesn't touch:
https://github.com/scipy/scipy/blob/b8c4794bf480f9f37361eda1abc8ba4a57b086e8/scipy/spatial/distance.py#L1799-L1807

I don't think there is a benchmark for the non-batched `spatial.distance.*` versions which is why it doesn't show up in the benchmarks.